requirements.txt
remla25-team7-lib-ml
scikit-learn
pandas
joblib
requests
dvc
dvc[gdrive]
pytest
pylint
coverage 
coverage-badge